Add concept of a repository and environment
The repository returns a GitRepo instance that can be used to easily fire off commands to the underlying git repository. This is not currently used in the Gitify core, but is used in GitifyWatch and can be convenient for further extensions.

The environment is a list of options for a specific host, or server. It works based off the domain. Not yet used in the Gitify core, but will be and is also used in GitifyWatch.