Add ability to install local packages
When using the --local flag, any packages inside the /core/packages folder that are not present in the `transport_packages` table will be installed. If the package is already installed, but it's an older version, the package will be updated.

This code is taken from the MODX core. It's similar to using the "Search for local packages" option in the package installer.

It can be useful for:

- Speeding up installations by not having to download every package first
- Keeping package versions in sync with a donor install
- Installing private packages outside of any package provider
- Testing new package versions during local development

Biggest drawback is that any package installed this way will have its provider listed as "Local", meaning it will not alert you of available package updates from the MODX repository.