Use values from external XML file first, ask questions later
The createMODXConfig function now evaluates any provided config data first, and only becomes interactive if a certain value is not provided.

This way, it integrates nicely with the ability to move the core. If you specify an alternative location in config.xml, the core will be moved there.

The Core Path and Manager Directory questions are also shortened to match the formatting of the others.

I'm also scratching my own itch by adding variables for DB charset and collation. See: https://github.com/modmore/Gitify/issues/265